refactoring middlewares
