--- services/conference/src/scripts/models/TrackMiddleware/index.ts
@@ -1,2 +1,0 @@
-import './MuteLocal'
-import './ChooseDevice'
