+++ services/conference/src/scripts/models/middleware/TrackMiddleware/index.ts
@@ -1,0 +1,2 @@
+import './MuteLocal'
+import './ChooseDevice'
