+++ services/conference/src/scripts/models/middleware/initStores/localParticipant.ts
